+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - added loading projects overlay instead of showing Open Project page while loading projects
 - added additional search options to List Tasks tool
 - added option to autostart AiderDesk server on run command
+- correctly calculating cached tokens for Requesty provider models
 
 ## [0.76.0]